third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger-pause/pause-in-internal-script-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that internal scripts unknown to front-end are processed correctly when appear in debugger call frames. Bug 64995


Running: testSetBreakpoint
Script source was shown.
Set timer for test function.
Script execution paused.
Call stack:
    0)  (pause-in-internal-script.js:22)
    1) testFunction (pause-in-internal-script.js:20)
Script execution resumed.
